core: fix warnings surfaced by rustc 1.98
Newer stable rustc grew four new complaints about core, and its sharper
unused-import analysis exposed two long-standing warnings that polluted
every cargo check run. All of them become hard errors under CI clippy's
--deny=warnings once the toolchain moves, so fix them ahead of the bump:

- ExternalAggState derives PartialEq over function pointers on purpose
  (same registered callbacks = same aggregate), so allow
  unpredictable_function_pointer_comparisons there.
- vector_from_text now spells out that the returned Vector borrows its
  text argument.
- The min_specialization impls move to their own file included behind
  cfg(nightly): 'default fn' is gated syntax and stable rustc warns
  about it before cfg stripping, so it must not be parsed at all.
- The raw proto re-exports in the logical log and the vdbe CollationSeq
  re-export only have consumers behind conn_raw_api and bench/fuzz
  respectively, so gate them the same way instead of exporting them
  unconditionally.

Tests: cargo check -p turso_core emits zero warnings on 1.88 and 1.98
for default, conn_raw_api, and bench feature sets; cargo fmt --check
clean on both.
