+++ core/alloc/collections/iterator.rs
@@ -1,11 +1,11 @@
 use super::{TursoFromIterator, TursoIteratorExt};
 use crate::alloc::TryReserveError;
+
 #[cfg(nightly)]
-use {
-    super::{trusted_len, TursoTryWithCapacityExt},
-    crate::alloc::Vec,
-    std::iter::TrustedLen,
-};
+#[path = "iterator_spec.rs"]
+mod spec;
+#[cfg(nightly)]
+use spec::TrySpecFromResultIter;
 
 fn try_from_result_iter<T, E, F, C, I>(iter: I) -> Result<Result<C, F>, TryReserveError>
 where
@@ -24,47 +24,6 @@ where
     if let Some(error) = error {
         Ok(Err(error))
     } else {
-        Ok(Ok(values))
-    }
-}
-
-#[cfg(nightly)]
-trait TrySpecFromResultIter<T, E, I>: Sized {
-    fn try_from_result_iter(iter: I) -> Result<Self, TryReserveError>;
-}
-
-#[cfg(nightly)]
-impl<T, E, F, C, I> TrySpecFromResultIter<T, E, I> for Result<C, F>
-where
-    C: TursoFromIterator<T>,
-    F: From<E>,
-    I: Iterator<Item = Result<T, E>>,
-{
-    default fn try_from_result_iter(iter: I) -> Result<Self, TryReserveError> {
-        try_from_result_iter(iter)
-    }
-}
-
-#[cfg(nightly)]
-impl<T, E, F, I> TrySpecFromResultIter<T, E, I> for Result<Vec<T>, F>
-where
-    F: From<E>,
-    I: TrustedLen<Item = Result<T, E>>,
-{
-    fn try_from_result_iter(iter: I) -> Result<Self, TryReserveError> {
-        // Specialize before the fallback's `map_while`, which cannot retain
-        // `TrustedLen` because an error may stop collection early.
-        let additional = trusted_len(iter.size_hint())?;
-        let mut values = <Vec<T> as TursoTryWithCapacityExt>::try_with_capacity_ext(additional)?;
-
-        for item in iter {
-            match item {
-                Ok(value) => {
-                    let _ = values.push_within_capacity(value);
-                }
-                Err(error) => return Ok(Err(F::from(error))),
-            }
-        }
         Ok(Ok(values))
     }
 }
